+++ backend/src/main/java/com/kongdak/controller/dto/request/CoupleConnectRequest.java
@@ -1,4 +1,4 @@
-package com.kongdak.controller.dto;
+package com.kongdak.controller.dto.request;
 
 import java.time.LocalDateTime;
