+++ backend/src/main/java/com/kongdak/controller/dto/request/ScheduleCreateRequest.java
@@ -1,4 +1,4 @@
-package com.kongdak.controller.dto;
+package com.kongdak.controller.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.kongdak.domain.calendar.ScheduleCategory;
